- Add CUDA kernels for rotary_pos_embedding and gemma_rotary_embedding
- Add Triton ops for matmul, norm, and position embedding
- Enable bf16, torch.compile, and CUDA graph support for Eagle backbone
- Update gr00t libero10 config (state_dim=64, lr adjustment)
- Add unit tests for gr00t VLA model
- Add C++/CUDA code style configs (.clang-format, .clang-tidy, .cpplint)

+++ fluxvla/models/vlas/llava_vla.py
@@ -199,7 +199,7 @@ class LlavaVLA(OpenVLA):
                     'Output must contain either hidden_states or last_hidden_state.'  # noqa: E501
                 last_hidden_state = output['last_hidden_state']
         pred_actions = self.vla_head.predict_action(
-            input_features=last_hidden_state.float(),
+            input_features=last_hidden_state,
             states=states,
             attention_mask=fused_attention_mask,
             embodiment_ids=embodiment_ids)
